moved to storage subcommand
